requirements.txt
# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
dash==2.5.1
gunicorn==20.1.0
numpy==1.23.1
django==3.2.5
django-environ==0.9.0
dj-database-url==0.5.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
django-apscheduler==0.6.2